follow up to #2392 refactor a ci test that was calling internet
