+++ karate-core/src/test/java/com/intuit/karate/http/mock.feature
@@ -1,0 +1,5 @@
+@ignore
+Feature:
+
+Scenario:
+* def response = ({ body: null, path: requestPath })
